+++ src/network_lib/src/jaln_session.h
@@ -101,7 +101,7 @@ struct jaln_sub_data {
  */
 struct jaln_pub_data {
 	struct jaln_payload_feeder journal_feeder;  //!< the jaln_payload_feeder for sending a journal record.
-	int vortex_feeder_sz;                       //!< The size (as reported to the Vortex engine) of this message.
+	int64_t vortex_feeder_sz;                       //!< The size (as reported to the Vortex engine) of this message.
 	int msg_no;                                 //!< The message number we are replying to
 
 	char *nonce;                            //!< The nonce of the last record sent.
